Added binding and validation for race & team forms

+++ src/main/java/com/marathonfront/domain/Race.java
@@ -9,6 +9,6 @@ public class Race {
 
     private long id;
     private String name;
-    private long distance;
+    private Integer distance;
     private BigDecimal price;
 }
